+++ frontend/src/constants/constants.js
@@ -31,47 +31,6 @@ export const COLORS = {
 
 export const HIT_PIXEL_TO_TOLERANCE = 3
 
-export const BACKOFFICE_SEARCH_PROPERTIES = ['layerName', 'zone', 'region', 'regulatoryReferences']
-
-export const FRENCH_REGION_LIST = [
-  'Auvergne-Rhône-Alpes',
-  'Bourgogne-Franche-Comté',
-  'Bretagne',
-  'Centre-Val de Loire',
-  'Corse',
-  'Grand Est',
-  'Hauts-de-France',
-  'Ile-de-France',
-  'Normandie',
-  'Nouvelle-Aquitaine',
-  'Occitanie',
-  'Pays de la Loire',
-  'Provence-Alpes-Côte d’Azur',
-  'Guadeloupe',
-  'Martinique',
-  'Guyanne',
-  'La Réunion',
-  'Mayotte'
-]
-
-export const INFO_TEXT = {
-  layerNamePart1: 'Avant de créer une nouvelle thématique, vérifiez bien qu\'il n\'en existe pas déjà une qui pourrait correspondre.',
-  editLayerNamePart1: 'Attention, le nom de la thématique sera modifié pour toutes les zones qu\'elle contient.',
-  layerNamePart2: `
-  Le nom de la thématique doit permettre de connaître en un coup d'œil le lieu et le sujet de la réglementation.
-  Il peut être intéressant de mentionner une zone totalement interdite à la pêche avec le mot "interdiction" (dans le champ "autres indications").
-  
-  NB : le lieu indiqué peut être une façade maritime (NAMO), une région (Bretagne), un département (Vendée), une commune (Saint-Malo), une zone géographique (Cotentin), une mer (Mer Baltique), etc.`,
-  zoneName: `Comme pour les thématiques, le nom des zones doit être aussi explicite que possible. Le couple thématique / zone fonctionne comme un tout, qui permet à l'utilisateur de comprendre rapidement quelle réglementation il consulte.
-  
-  Le nom de la zone peut être
-  - un nom géographique ("Ile-Rousse", ou "Bande des 3 miles"),
-  - ou numérique ("Zone 1"),
-  - ou spécifiant une autre caractéristique ("Zone pour navire > 10m")`
-}
-
-export const INFINITE = 'infinite'
-
 /**
  * @enum {string}
  */
