+++ frontend/src/domain/use_cases/searchRegulatoryLayers.js
@@ -33,7 +33,7 @@ const searchRegulatoryLayers = (searchFields, inputsAreEmpty) => {
     if (extent?.length === 4) {
       return getRegulatoryZonesInExtentFromAPI(extent, getState().global.inBackofficeMode)
         .then(features => worker.convertGeoJSONFeaturesToStructuredRegulatoryObject(features))
-        .then(response => getRegulatoryLayersWithoutTerritory(response))
+        .then(regulatoryLayers => getRegulatoryLayersWithoutTerritory(regulatoryLayers))
         .then(filteredRegulatoryLayers => {
           if (inputsAreEmpty) {
             return filteredRegulatoryLayers
