+++ frontend/src/features/backoffice/create_regulation/regulatory_identification/RegulationRegionLine.js
@@ -3,7 +3,7 @@ import { ContentLine } from '../../../commonStyles/Backoffice.style'
 import { Label } from '../../../commonStyles/Input.style'
 import CustomSelectComponent from '../custom_form/CustomSelectComponent'
 import { formatDataForSelectPicker } from '../../../../utils'
-import { FRENCH_REGION_LIST } from '../../../../constants/constants'
+import { FRENCH_REGION_LIST } from '../../constants'
 import Tag from '../Tag'
 import MenuItem from '../custom_form/MenuItem'
